Make range proof accept Scalar input values
Most of the required changes for this consist of simply replacing the u64 type
with Scalar in many places within the code, with one significant exception: I
had to implement the extraction of a single bit from a Scalar.

Extracting a bit from a Scalar is done by indexing into the Scalar's underlying
array of bytes, and then shifting and masking the appropriate byte. In both
cases where this operation is used, the index is not secret (it just iterates
from 0 to the range proof's bitsize), so the timing and memory access patterns
of this operation shouldn't enable any side-channel attacks.

I was also concerned that the bit extraction function might save the indexed
byte and/or the extracted bit into some temporary variables on the stack, thus
potentially exposing them to memory dump attacks, etc. However, a look at the
generated assembly (compiled on x86_64 with optimisations) confirms that the
indexed byte is read directly into a register, and that the extraction of the
requested bit from said byte is done entirely within registers. Then, once the
extraction of the bit is done, the bit's value is processed exactly as it was
before this commit. Funnily enough, this includes converting the bit into a
temporary Scalar value, which is definitely saved on the stack and is never
explicitly cleared, so maybe I'm just worrying too much about these issues...